+++ lib/src/analyzers/lint_analyzer/rules/rules_list/prefer_single_widget_per_file/visitor.dart
@@ -3,7 +3,8 @@ part of 'prefer_single_widget_per_file.dart';
 class _Visitor extends SimpleAstVisitor<void> {
   final _nodes = <ClassDeclaration>[];
 
-  Iterable<ClassDeclaration> get nodes => _nodes;
+  Iterable<ClassDeclaration> get nodes =>
+      _nodes.length > 1 ? _nodes.skip(1) : [];
 
   @override
   void visitClassDeclaration(ClassDeclaration node) {