+++ test/analyzers/lint_analyzer/rules/rules_list/prefer_single_widget_per_file/examples/correct_stateless_widget_example.dart
@@ -1,6 +1,6 @@
 import 'flutter_defines.dart';
 
-class someWidget extends StatelessWidget {
+class SomeWidget extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
 //    ...